+++ lib/enough_mail_flutter.dart
@@ -1,3 +1,5 @@
+/// UI components useful for displaying mails
 library enough_mail_flutter;
 
 export 'mime_message_viewer.dart';
+export 'mime_message_downloader.dart';