+++ test/enough_mail_flutter_test.dart
@@ -2,6 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:enough_mail_flutter/enough_mail_flutter.dart';
 
+//TODO create tests
 void main() {
   test('adds one to input values', () {
     // final calculator = Calculator();