+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.0] - 2020-12-04.
+* Added `MimeMessageDownloader` widget.
+* Improve documentation.
+
 ## [0.1.0] - 2020-12-04.
 
-* Initial release with the `MimeMessageViewer` component.
+* Initial release with the `MimeMessageViewer` widget.